List of changes:
- Introduce OrderPlacer
- FIx horizontal shift on profile page when modal is opened
