+++ app/(root)/shop/_components/ProductPreview/index.ts
@@ -1,0 +1,3 @@
+import ProductPreview from "./ProductPreview";
+
+export {ProductPreview};
