+++ app/(root)/shop/_page.tsx
@@ -2,55 +2,17 @@
 
 import { useState } from 'react';
 import { Product as ProductType } from '@/types/product'
-import { useProducts } from '@/context/products';
+import { useProductsContext } from '@/context/products';
 import { Modal } from "@/components/Modal";
-import { Product, Placeholder } from './_components/Product'
-import ProductPreview from './_components/ProductPreview'
-import Listing from './_components/Listing';
+import { ProductPreview } from './_components/ProductPreview'
+import { ProductList, PlaceholderList } from './_components/ProductList';
 
-interface ProductListProps { 
-    products: ProductType[], 
-    searchPrompt: string, 
-    setModalProduct: (product: ProductType) => void 
-};
 
-const ProductList = ({ products, searchPrompt, setModalProduct }: ProductListProps) => {
-    const trimmedPrompt = searchPrompt.trim();
-    const filteredProducts = products.filter((product) => {
-        if (trimmedPrompt.length == 0) return true;
-
-        if (product.name.toLowerCase().includes(trimmedPrompt.toLowerCase()))
-            return true;
-
-        if (product._hasVariants) {
-            for (const variant of product.variants) {
-                if (variant.name.toLowerCase().includes(trimmedPrompt.toLowerCase()))
-                    return true;
-            }
-        }
-        return false;
-    });
-
-    return (
-        filteredProducts.length == 0 ? (
-            <h1 className='font-semibold'>No products found for "{searchPrompt}".</h1>
-        ) : (
-            <Listing products={filteredProducts} callback={(product: ProductType) => (<Product product={product} key={product._id} onClick={() => setModalProduct(product)} />)} />
-        )
-    )
-}
-
-const PlaceholderList = () => {
-    return (
-        <Listing products={Array(12).fill(undefined)} callback={(_, i) => (<Placeholder key={i} />)} />
-    )
-};
-
-export default function ShopPage_ () {
+export default function ShopPage_() {
     const [modalProduct, setModalProduct] = useState<ProductType | null>(null);
     const [searchPrompt, setSearchPrompt] = useState<string>("");
 
-    const { products, loading, error } = useProducts();
+    const { products, loading, error } = useProductsContext();
 
     return (
         <>
