+++ app/api/order/cake/create/route.ts
@@ -1,10 +1,11 @@
 import { NextRequest, NextResponse } from "next/server";
-import { CakeOrder, createNewCakeOrder } from "../_lib/cakeorders";
+import { Cakes } from "@/libs/orders";
+import { CakeOrder } from "@/types/orders";
 
 export const dynamic = 'force-dynamic'
 
 export async function POST(request: NextRequest) {
     const data: CakeOrder = await request.json();
-    const id = createNewCakeOrder(data);
+    const id = Cakes.placeOrder(data);
     return NextResponse.json({ ...data, id });
 }
