+++ app/api/order/cake/route.ts
@@ -1,10 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
-import { getCakeOrder } from "./_lib/cakeorders";
-import { UUID } from "crypto";
+import { Cakes } from "@/libs/orders";
 
 export const dynamic = 'force-dynamic'
 
 export async function GET(request: NextRequest) {
     const searchParams = request.nextUrl.searchParams;
-    return NextResponse.json(getCakeOrder(searchParams.get("id")!));
+    return NextResponse.json(Cakes.getOrder(searchParams.get("id")!));
 }
