+++ app/api/order/products/route.ts
@@ -1,9 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
-import { getProductOrder } from "./_lib/products";
+import {Products} from '@/libs/orders'
 
 export const dynamic = 'force-dynamic'
 
 export async function GET(request: NextRequest) {
     const searchParams = request.nextUrl.searchParams;
-    return NextResponse.json(getProductOrder(searchParams.get("id")!));
+    return NextResponse.json(Products.getOrder(searchParams.get("id")!));
 }
