--- app/api/products/create/route.ts
@@ -1,8 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-export const dynamic = 'force-dynamic'
-
-export async function POST(request: NextRequest) {
-    const data = await request.json();
-    return NextResponse.json(data);
-}
