+++ libs/orders.ts
@@ -1,0 +1,5 @@
+import { CakeOrder, ProductOrder } from "@/types/orders";
+import { OrderPlacer } from "./orderplacer";
+
+export const Products = new OrderPlacer<ProductOrder>("products");
+export const Cakes = new OrderPlacer<CakeOrder>("cakes");
