+++ libs/products.ts
